+++ server/package.json
@@ -1,8 +1,9 @@
 {
-  "name": "netflix-clone",
+  "name": "server",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "dev": "nodemon index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -11,22 +12,15 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.1.0",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
-    "ejs": "^3.1.9",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "express-flash": "^0.0.2",
-    "express-session": "^1.17.3",
-    "html": "^1.0.0",
-    "mongodb": "^5.5.0",
-    "mongoose": "^7.2.2",
-    "node-fetch": "^2.6.11",
-    "nodemon": "^2.0.22",
-    "passport": "^0.6.0",
-    "passport-local": "^1.0.0"
-  },
-  "devDependencies": {
-    "dotenv": "^16.1.4"
+    "jsonwebtoken": "^9.0.0",
+    "mongodb": "^6.2.0",
+    "mongoose": "^6.1.7",
+    "multer": "^1.4.5-lts.1",
+    "nodemon": "^2.0.22"
   }
 }
