--- server/views/homepage.html
@@ -1,154 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../client/home.css">
-    <title>Movies</title>
-</head>
-<body>
-    <div class="parent">
-        <div class="child">
-            <div class="nav">
-                <div class="nav-images">
-                <img id="menu" src="../client/images/icon-menu.svg" alt="">
-                <img id="logo" src="../client/images/trailer-logo.jpg" alt="">
-            </div>
-                <div class="search">
-                    <input type="text" placeholder="Search">
-                    <img src="../client/images/icons8-search.svg" alt="">
-                </div>
-            </div>
-            <div class="highlights">
-
-            </div>
-            <div class="trending">
-                <div class="head">
-                    <h1>Trending</h1>
-
-                </div>
-                <div class="today">
-                    <div class="data">
-                        <h1>Today</h1>
-                        <div class="box">
-                            
-                        </div>
-                        <div class="description">
-                            <h3>Movie Name</h3>
-                            <h3>DAte</h3>
-                        </div>
-                    </div>
-                </div>
-                <div class="thisweek">
-                    <div class="head"><h1>This Week</h1></div>
-                    <div class="data">
-                        <div class="box">
-                            
-                        </div>
-                        <div class="description">
-                            <h3>Movie Name</h3>
-                            <h3>DAte</h3>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="popular">
-                <div class="buttons">
-                    <h2>Everyone's watching</h2>
-                <button onclick="showContent('streaming')">Streaming</button>
-                <button onclick="showContent('tv')">On Tv</button>
-                <button onclick="showContent('theatres')">In theatres</button>
-            </div>
-
-                <div id="streaming" class="content">
-                    <div class="data">
-                        <div class="box">
-    
-                        </div>
-                        <div class="description">
-                            <h3>Movie Name</h3>
-                            <h3>DAte</h3>
-                        </div>
-                    </div>
-                </div>
-                <div id="tv" class="content">
-                    <div class="data">
-                        <div class="box">
-    
-                        </div>
-                        <div class="description">
-                            <h3>Movie Name</h3>
-                            <h3>DAte</h3>
-                        </div>
-                    </div>
-                </div>
-                <div id="theatres" class="content">
-                    <div class="data">
-                        <div class="box">
-    
-                        </div>
-                        <div class="description">
-                            <h3>Movie Name</h3>
-                            <h3>DAte</h3>
-                        </div>
-                    </div>
-                </div>
-
-            </div>
-            <div class="genres">
-                <div class="toprated">
-                    <h1>Top Rated</h1>
-                    <div class="data">
-                    <div class="box">
-
-                    </div>
-                    <div class="description">
-                        <h3>Movie Name</h3>
-                        <h3>DAte</h3>
-                    </div>
-                </div>
-                </div>
-                <div class="comedy">
-                    <h1>Comedy</h1>
-                    <div class="data">
-                    <div class="box">
-                        
-                    </div>
-                    <div class="description">
-                        <h3>Movie Name</h3>
-                        <h3>DAte</h3>
-                    </div>
-                </div>
-                </div>
-                <div class="horror">
-                    <h1>Horror</h1>
-                    <div class="data">
-                    <div class="box">
-                        
-                    </div>
-                    <div class="description">
-                        <h3>Movie Name</h3>
-                        <h3>DAte</h3>
-                    </div>
-                </div>
-                </div>
-                <div class="docu">
-                    <h1>Documentaries</h1>
-                    <div class="data">
-                    <div class="box">
-                        
-                    </div>
-                    <div class="description">
-                        <h3>Movie Name</h3>
-                        <h3>DAte</h3>
-                    </div>
-                </div>
-                </div>
-            </div>
-            <div class="footer"></div>
-        </div>
-    </div>
-    <script src="./index.js"></script>
-</body>
-</html>
